+++ test/unit/fuzz-tools.test.js
@@ -301,6 +301,7 @@ import * as hincheyV321 from '../../lib/hinchey-v321.js';
 import * as biRadsV322 from '../../lib/bi-rads-v322.js';
 import * as siewertV323 from '../../lib/siewert-v323.js';
 import * as wexnerV324 from '../../lib/wexner-v324.js';
+import * as lungRadsV325 from '../../lib/lung-rads-v325.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -547,6 +548,7 @@ const MODULES = {
   'bi-rads-v322.js': biRadsV322,
   'siewert-v323.js': siewertV323,
   'wexner-v324.js': wexnerV324,
+  'lung-rads-v325.js': lungRadsV325,
   'tb-testing.js': tbTesting,
 };